+++ src/schemathesis/exceptions.py
@@ -1,2 +1,2 @@
-class InvalidEndpoint(Exception):
-    pass
+class InvalidSchema(Exception):
+    """Schema associated with an endpoint contains an error."""