+++ src/schemathesis/models.py
@@ -84,7 +84,7 @@ class Endpoint:
     body: Body = attr.ib(factory=empty_object)  # pragma: no mutate
     form_data: FormData = attr.ib(factory=empty_object)  # pragma: no mutate
 
-    def as_strategy(self) -> Optional[SearchStrategy]:
+    def as_strategy(self) -> SearchStrategy:
         from ._hypothesis import get_case_strategy  # pylint: disable=import-outside-toplevel
 
         return get_case_strategy(self)